+++ src/utils/mod.rs
@@ -1,3 +1,3 @@
+pub mod error_utils;
 pub mod hkt_utils;
 pub mod transform_utils;
-pub mod error_utils;